backend/requirements.txt
Django==4.1.3
Pillow==9.5.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
requests==2.30.0
oauthlib==3.2.2
python-dotenv==1.0.0
requests==2.30.0
social-auth-app-django==5.2.0
requests-oauthlib==1.3.1
